Delete Permissions

+++ backend/src/resolvers/Mutation.js
@@ -1,14 +1,15 @@
-const bcrypt = require("bcryptjs");
-const jwt = require("jsonwebtoken");
-const { randomBytes } = require("crypto");
-const { promisify } = require("util");
-const { transport, makeANiceEmail } = require("../mail");
+const bcrypt = require('bcryptjs');
+const jwt = require('jsonwebtoken');
+const { hasPermission } = require('../utils');
+const { randomBytes } = require('crypto');
+const { promisify } = require('util');
+const { transport, makeANiceEmail } = require('../mail');
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
     //TODO Check if they are logged in
     if (!ctx.request.userId) {
-      throw new Error("You must be logged in to do that!");
+      throw new Error('You must be logged in to do that!');
     }
 
     const item = await ctx.db.mutation.createItem(
@@ -48,9 +49,21 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // Find the item
-    const item = await ctx.db.query.item({ where }, `{ id, title }`);
+    const item = await ctx.db.query.item(
+      { where },
+      `{ id, title, user { id } }`
+    );
     // Check if they own that item or have permissions
-    //TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    ); // some = atleast 1 is true
+    // ['DOG', 'CAT'].some(pet => ['BIRD', ['SNAKE'].includes(pet)); == false
+    // ['DOG', 'CAT'].some(pet => ['BIRD', 'SNAKE', 'CAT'].includes(pet)); == true
+    if (!ownsItem && hasPermissions) {
+      throw new Error("You don't have permission to do that!");
+    }
     // Delete it
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -66,7 +79,7 @@ const Mutations = {
         data: {
           ...args,
           password,
-          permissions: { set: ["USER"] }
+          permissions: { set: ['USER'] }
         }
       },
       info
@@ -74,7 +87,7 @@ const Mutations = {
     // Create JWT token for them
     const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
     // set JWT as a cookie on response
-    ctx.response.cookie("token", token, {
+    ctx.response.cookie('token', token, {
       httpOnly: true,
       maxAge: 1000 * 60 * 60 * 24 * 365 // 1 year cookie)
     });
@@ -95,7 +108,7 @@ const Mutations = {
     const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
 
     // set cookie with token
-    ctx.response.cookie("token", token, {
+    ctx.response.cookie('token', token, {
       httpOnly: true,
       maxAge: 1000 * 60 * 60 * 24 * 365
     });
@@ -103,8 +116,8 @@ const Mutations = {
     return user;
   },
   signout(parent, args, ctx, info) {
-    ctx.response.clearCookie("token");
-    return { message: "Goodbye!" };
+    ctx.response.clearCookie('token');
+    return { message: 'Goodbye!' };
   },
   async requestReset(parent, args, ctx, info) {
     // Check if it's a real user
@@ -114,7 +127,7 @@ const Mutations = {
     }
     // set a reset token expiry on that user
     const randomBytesPromisifed = promisify(randomBytes);
-    const resetToken = (await randomBytesPromisifed(20)).toString("hex");
+    const resetToken = (await randomBytesPromisifed(20)).toString('hex');
     const resetTokenExpiry = Date.now() + 3600000; // 1 hour from now
     const res = await ctx.db.mutation.updateUser({
       where: { email: args.email },
@@ -123,20 +136,20 @@ const Mutations = {
     // console.log(res);
     // email them a reset token
     const mailRes = await transport.sendMail({
-      from: "allanmolson@gmail.com",
+      from: 'allanmolson@gmail.com',
       to: user.email,
-      subject: "Your Password Reset Token",
+      subject: 'Your Password Reset Token',
       html: makeANiceEmail(`Your Password Reset Token is Here! 
       \n\n 
       <a href="${process.env.FRONTEND_URL}/reset?resetToken=${resetToken}">Click here to Reset</a>`)
     });
     //  return the message
-    return { message: "Thanks!" };
+    return { message: 'Thanks!' };
   },
   async resetPassword(parent, args, ctx, info) {
     // check if passwords match
     if (args.password !== args.confirmPassword) {
-      throw new Error("Your Passwords dont match!");
+      throw new Error('Your Passwords dont match!');
     }
     //check if it's a ligit reset token
 
@@ -148,7 +161,7 @@ const Mutations = {
       }
     });
     if (!user) {
-      throw new Error("This token is either invalid or expired");
+      throw new Error('This token is either invalid or expired');
     }
     // hash their new password
     const password = await bcrypt.hash(args.password, 10);
@@ -167,12 +180,43 @@ const Mutations = {
     // generate JWT
     const token = jwt.sign({ userId: updatedUser.id }, process.env.APP_SECRET);
     // Set the JWT cookie
-    ctx.response.cookie("token", token, {
+    ctx.response.cookie('token', token, {
       httpOnly: true,
       maxAge: 1000 * 60 * 60 * 24 * 365
     });
     // return the new user
     return updatedUser;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    // checked if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in');
+    }
+    // query current user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId
+        }
+      },
+      info
+    );
+    // check if they have permissions to do this
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+    // update the permissions
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions
+          }
+        },
+        where: {
+          id: args.userId
+        }
+      },
+      info
+    );
   }
 };
